requirements.txt
python-dotenv
streamlit
streamlit-extras
wikipedia
mwparserfromhell
diskcache
snowflake